Rename and move constants/discord.ts to utils/constants.ts
