+++ src/module/requestManager.ts
@@ -1,8 +1,8 @@
 import { delay } from "../../deps.ts";
-import { baseEndpoints } from "../constants/discord.ts";
 import { HttpResponseCode } from "../types/discord.ts";
 import { Errors } from "../types/errors.ts";
 import { RequestMethods } from "../types/fetch.ts";
+import { baseEndpoints } from "../utils/constants.ts";
 import { authorization, eventHandlers } from "./client.ts";
 
 const pathQueues: { [key: string]: QueuedRequest[] } = {};
